+++ myUtils/auth.py
@@ -15,6 +15,7 @@ from uploader.smzdm_uploader.main import cookie_auth as smzdm_cookie_auth
 from uploader.toutiao_uploader.main import cookie_auth as toutiao_cookie_auth
 from uploader.ctrip_uploader.main import cookie_auth as ctrip_cookie_auth
 from uploader.sohu_uploader.main import cookie_auth as sohu_cookie_auth
+from uploader.weibo_uploader.main import cookie_auth as weibo_cookie_auth
 
 
 async def cookie_auth_douyin(account_file):
@@ -137,6 +138,9 @@ async def check_cookie(type, file_path):
         # 搜狐号
         case 9:
             return await sohu_cookie_auth(str(cookie_path))
+        # 微博
+        case 10:
+            return await weibo_cookie_auth(str(cookie_path))
         case _:
             return False